Create Excel_Sheet.xlsx
This is the Excel Sheet from which the main Python logic will take values and search the value's original website to find the price. The Python file will contain the logic to write these prices to an HTML file for easy transferral (with more Python logic or manual work) to this file so that you can have the value's current price.