Create Price_Scraper.py
This file contains the primary logic of taking the Excel Spreadsheet's values (or at least the ones given) and then searching for them using a static URL address. After the specific field is found, it writes the field into an HTML document for the user to then personally add to their original spreadsheet. Python code to add the new/updated field has not been built yet due to time constraints, but the basic building blocks are completed.